% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_h_index}
\alias{calculate_h_index}
\title{Calculate h-index}
\usage{
calculate_h_index(citations)
}
\arguments{
\item{citations}{Numeric vector of citation counts}
}
\value{
Numeric h-index value
}
\description{
Calculate h-index from citation counts
}
