% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_i10_index}
\alias{calculate_i10_index}
\title{Calculate i10-index}
\usage{
calculate_i10_index(citations)
}
\arguments{
\item{citations}{Numeric vector of citation counts}
}
\value{
Numeric i10-index value
}
\description{
Calculate i10-index (number of papers with at least 10 citations)
}
