% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_m_index}
\alias{calculate_m_index}
\title{Calculate m-index}
\usage{
calculate_m_index(h_index, first_publication_year)
}
\arguments{
\item{h_index}{Numeric h-index value}

\item{first_publication_year}{Numeric year of first publication}
}
\value{
Numeric m-index value
}
\description{
Calculate m-index (h-index divided by number of years since first publication)
}
