% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{calculateMetrics}
\alias{calculateMetrics}
\title{Calculate Citation Metrics}
\usage{
calculateMetrics(object, ...)
}
\arguments{
\item{object}{An object of class `ScholarProfile`.}

\item{metric}{A character string specifying the metric to calculate. Options are: "h_index", "i10_index", or "m_index".}
}
\value{
A numeric value representing the chosen citation metric.
}
\description{
Calculates various citation metrics (e.g., h-index, i10-index, m-index) for a scholar's profile.
}
\examples{
scholar_profile <- new("ScholarProfile", publications = data.frame(citedby = c(10, 20, 30, 40), year = c(2000, 2005, 2010, 2015)))
calculateMetrics(scholar_profile, metric = "h_index")
calculateMetrics(scholar_profile, metric = "i10_index")
calculateMetrics(scholar_profile, metric = "m_index")

}
