% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{CompareScholars}
\alias{CompareScholars}
\title{Compare Scholars}
\usage{
CompareScholars(
  scholar_profiles,
  metrics = c("h_index", "i10_index", "citations_total", "publications_count")
)
}
\arguments{
\item{scholar_profiles}{list of ScholarProfile objects}

\item{metrics}{Character vector of metrics to compare}
}
\value{
data.frame with comparison results
}
\description{
Compare multiple scholars based on various metrics
}
