% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{CreateCitationNetwork}
\alias{CreateCitationNetwork}
\title{Create Citation Network}
\usage{
CreateCitationNetwork(scholar_profiles, min_citations = 5)
}
\arguments{
\item{scholar_profiles}{list of ScholarProfile objects}

\item{min_citations}{Integer minimum citations for a paper to be included}
}
\value{
igraph object representing the citation network
}
\description{
Create a co-citation network from scholar publications
}
