% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{estimate_impact_factor}
\alias{estimate_impact_factor}
\title{Estimate Impact Factor}
\usage{
estimate_impact_factor(publications, year_range = 5)
}
\arguments{
\item{publications}{data.frame containing publication data}

\item{year_range}{Integer number of years to consider (default: 5)}
}
\value{
data.frame with estimated impact factors by journal
}
\description{
Estimate journal impact factor based on citation data
}
