% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_co_citation_network}
\alias{get_co_citation_network}
\title{Get Co-citation Network}
\usage{
get_co_citation_network(publication_ids, max_connections = 100)
}
\arguments{
\item{publication_ids}{Character vector of publication IDs}

\item{max_connections}{Integer maximum number of connections to show}
}
\value{
igraph object representing co-citation network
}
\description{
Get co-citation network for a set of publications
}
