% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{ScholarData-class}
\alias{ScholarData-class}
\title{ScholarData S4 Class}
\description{
S4 class for storing comprehensive scholar data including citation history
}
\section{Slots}{

\describe{
\item{\code{profiles}}{list of ScholarProfile objects}

\item{\code{citation_history}}{data.frame containing historical citation data}

\item{\code{collaboration_network}}{igraph object representing collaboration network}

\item{\code{analysis_date}}{Date object when analysis was performed}
}}

