% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ScholarProfile}
\alias{ScholarProfile}
\title{Scholar Profile}
\usage{
ScholarProfile(scholar_id, max_publications = 100, include_network = TRUE)
}
\arguments{
\item{scholar_id}{Character string Google Scholar ID}

\item{max_publications}{Integer maximum number of publications to retrieve}

\item{include_network}{Logical whether to include collaboration network analysis}
}
\value{
ScholarData object containing complete analysis
}
\description{
Create a comprehensive scholar profile with analysis
}
