% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{plot}
\alias{plot}
\alias{plot,ScholarProfile,ANY-method}
\title{Plot Scholar Profile Metrics}
\usage{
plot(x, y, ...)

\S4method{plot}{ScholarProfile,ANY}(x, y, type = c("citations", "publications", "h_index"), ...)
}
\description{
Plot various metrics (citations, publications, or h-index) for a scholar's profile.
}
